Offload notify_push from a thread
